book-formatter/requirements.txt
click
rich
pyyaml
jinja2
python-docx
ebooklib
Pillow
